requirements.txt
django
djangorestframework
django-environ
djangorestframework_simplejwt
faker
factory_boy
coreapi
pyyaml
uritemplate